refactor: Rename TransferAmount filename
